+++ adam/__init__.py
@@ -1,4 +1,3 @@
-from .core import Logger
 from .features.simulation import Simulation
 from .features.configurations import ConfigurationsManager
 from .features.map_maker import MapMaker